refactor: change to schema
