+++ test/api/database/operations/createMany.test.ts
@@ -1,7 +1,7 @@
 import knex, { Knex } from 'knex';
 import { createMany } from '../../../../src/api/database/operations/createMany.js';
 import { getSchemaOverview } from '../../../../src/api/database/overview.js';
-import { User } from '../../../../src/config/types.js';
+import { User } from '../../../../src/api/database/schemas.js';
 import { config } from '../../../config.js';
 import { testDatabases } from '../../../utilities/testDatabases.js';
