+++ test/api/database/operations/createOne.test.ts
@@ -1,9 +1,9 @@
 import knex, { Knex } from 'knex';
 import { createOne } from '../../../../src/api/database/operations/createOne.js';
 import { getSchemaOverview } from '../../../../src/api/database/overview.js';
-import { User } from '../../../../src/config/types.js';
 import { config } from '../../../config.js';
 import { testDatabases } from '../../../utilities/testDatabases.js';
+import { User } from '../../../../src/api/database/schemas.js';
 
 describe('Create One', () => {
   const tableName = 'superfast_users';
