+++ test/api/database/operations/deleteMany.test.ts
@@ -3,7 +3,7 @@ import { createMany } from '../../../../src/api/database/operations/createMany.j
 import { deleteMany } from '../../../../src/api/database/operations/deleteMany.js';
 import { readByQuery } from '../../../../src/api/database/operations/readByQuery.js';
 import { getSchemaOverview } from '../../../../src/api/database/overview.js';
-import { User } from '../../../../src/config/types.js';
+import { User } from '../../../../src/api/database/schemas.js';
 import { config } from '../../../config.js';
 import { testDatabases } from '../../../utilities/testDatabases.js';
