+++ test/api/database/operations/readById.test.ts
@@ -2,7 +2,7 @@ import knex, { Knex } from 'knex';
 import { createOne } from '../../../../src/api/database/operations/createOne.js';
 import { readById } from '../../../../src/api/database/operations/readById.js';
 import { getSchemaOverview } from '../../../../src/api/database/overview.js';
-import { User } from '../../../../src/config/types.js';
+import { User } from '../../../../src/api/database/schemas.js';
 import { config } from '../../../config.js';
 import { testDatabases } from '../../../utilities/testDatabases.js';
